Cleanup JNI handling and Android audio driver.
The API for accessing the java virtual machine and the jni environment
would not easily allow mixing with other jni using code. Now the Android
Audio driver exports a function that must be used to provide the jvm and
it holds a local pointer to the jvm for later usage.