Fix audio thread handling in the Android audio driver.
The shutdown function and the configure functions are not always called
in the same thread. Making sure that the thread is always
attached. Also, (maybe related with dash) the configure output function
can be called multiple times when playing a stream so making sure that
the resources are correctly freed up when the function is called again.